First stab at a script to build release demumble binaries for all platforms.
Still missing calls to `strip` on mac and linux, and doesn't zip up
binaries yet, but already produces binaries for linux/mac/windows.

Bump minimum cmake version to 3.2.0 to enable CMP0056 to on, and
remove "C" from project since 3.2.0 then tries to run the C compiler
in addition to the C++ compiler, and the new script only sets the C++
compiler.