dist.py: on linux, use CMAKE_SYSROOT instead of passing --sysroot in cflags
Close to no behavior change. It's "more correct" in that CMAKE_SYSROOT
also affects `find_*` commands.  Demumble doesn't use any of those, but
I suppose it's still nicer.

(This has the tiny side effect that the --sysroot flag is now on the
rule in CMakeFiles/rules.ninja instead of in each compile's cflags
in build.ninja.)